# Reworking view specs to be faster