--- a/spec/views/hyrax/base/show.json.jbuilder_spec.rb
+++ b/spec/views/hyrax/base/show.json.jbuilder_spec.rb
@@ -1,5 +1,5 @@
 RSpec.describe 'hyrax/base/show.json.jbuilder' do
-  let(:curation_concern) { create(:generic_work) }
+  let(:curation_concern) { build(:generic_work) }
 
   before do
     allow(curation_concern).to receive(:etag).and_return('W/"87f79d2244ded4239ad1f0e822c8429b1e72b66c"')