--- a/spec/views/hyrax/collections/_show_document_list.html.erb_spec.rb
+++ b/spec/views/hyrax/collections/_show_document_list.html.erb_spec.rb
@@ -1,28 +1,25 @@
 
 RSpec.describe 'hyrax/collections/_show_document_list.html.erb', type: :view do
-  let(:user) { create(:user) }
-  let(:collection) { mock_model(Collection) }
+  let(:documents) { ["Hello", "World"] }
 
-  let(:file) do
-    FileSet.create(creator: ["ggm"], title: ['One Hundred Years of Solitude']) do |fs|
-      fs.apply_depositor_metadata(user)
+  before do
+    stub_template('hyrax/collections/_show_document_list_row.html.erb' => "<%= document %>")
+  end
+  context 'when not logged in' do
+    it "renders the documents without an Action section" do
+      allow(view).to receive(:current_user).and_return(nil)
+      render('hyrax/collections/show_document_list.html.erb', documents: documents)
+      expect(rendered).to have_css('tbody', text: documents.join)
+      expect(rendered).not_to have_css('th', text: 'Action')
     end
   end
 
-  let(:documents) { [file] }
-
-  context 'when not logged in' do
-    before do
-      view.blacklight_config = Blacklight::Configuration.new
-      allow(view).to receive(:current_user).and_return(nil)
-      allow(file).to receive(:title_or_label).and_return("One Hundred Years of Solitude")
-      allow(file).to receive(:edit_people).and_return([])
-    end
-
-    it "renders collection" do
+  context 'when logged in' do
+    it "renders the documents with an Action section" do
+      allow(view).to receive(:current_user).and_return(true)
       render('hyrax/collections/show_document_list.html.erb', documents: documents)
-      expect(rendered).to have_content 'One Hundred Years of Solitude'
-      expect(rendered).not_to have_content 'Action'
+      expect(rendered).to have_css('tbody', text: documents.join)
+      expect(rendered).to have_css('th', text: 'Action')
     end
   end
 end