Do you want specs to run 10 seconds or so faster each time? I bet Travis will appreciate a few less cycles.

## Reworking view specs to not test called partials

@be56884d79888071793ed58b1704cfcc2cae7b7a

The spec was testing the implementation details of the subsequently
called (and tested) partial.

See `./spec/views/hyrax/collections/_show_document_list_row.html.erb_spec.rb`
for details of the spec.

With this change the individual tests go from:

* Before - Finished in 3.02 seconds (files took 6.82 seconds to load)
* After - Finished in 0.41196 seconds (files took 6.5 seconds to load)

Dropping 15 LDP requests

## Reworking view specs to not test called partials

@c036eb078dfae843942c13ac1b10d08ca8b759d0

The spec was testing the implementation details of the subsequently
called (and tested) partial.

See `./spec/views/hyrax/admin/admin_sets/_show_document_list_row.html.erb.erb_spec.rb`
for details of the spec.

With this change the individual tests go from:

* Before - Finished in 3.16 seconds (files took 6.85 seconds to load)
* After - Finished in 0.38931 seconds (files took 6.41 seconds to load)

Dropping 15 LDP requests

## Switching from create to build :generic_work

@aa8267ce1b160beab1b6041b74fc522922e3d86f

Before this change there are 15 LDP hits. After the change there are 3.
The single goes from 2.61 second to 0.7597.

## Reworking view specs to not test called partials

@7e955218e0a2e4757993632f90ed3ea53701ee8c

The spec was testing the implementation details of the subsequently
called (and tested) partial.

See `./spec/views/hyrax/base/_member.html.erb.erb_spec.rb`
for details of the spec.

With this change the individual tests go from:

* Before - Finished in 3.1 seconds (files took 6.73 seconds to load)
* After - Finished in 0.49883 seconds (files took 7.04 seconds to load)

Dropping 13 LDP requests